1.修复“首页账本图标和报表页面的账本图标大小不一致” 2.优化代码、消除警告

+++ SuiShouJi/Module/BookKeeping/BookKeepingHome/View/SSJHomeBudgetButton.m
@@ -98,7 +98,7 @@
 
     } else if ([model isKindOfClass:[SSJShareBookItem class]]){
         SSJShareBookItem *currentModel = (SSJShareBookItem *)model;
-        [self.button setTitle:[NSString stringWithFormat:@"%ld 人",currentModel.memberCount] forState:UIControlStateNormal];
+        [self.button setTitle:[NSString stringWithFormat:@"%ld 人", (long)currentModel.memberCount] forState:UIControlStateNormal];
     }
     [self sizeToFit];
     [self setNeedsLayout];
